backend/imaglee/pyproject.toml
[tool.poetry]
name = "imaglee_online"
version = "0.1.0"
description = "Website for Imaglee"
authors = ["Michal Krejci <mecumcrearum@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = ">=3.2,<4.0"
graphene-django = ">=2.15.0"
graphene = ">=2.1.8"
gunicorn = ">=20.0.4"
djangorestframework = ">=3.12.0"
django-filter = ">=2.4.0"
django-cors-headers = ">=3.5.0"
psycopg2-binary = ">=2.9.0"
redis = ">=4.0.0"
channels-redis = ">=3.2.0"
django-redis = ">=4.12.1"
channels = ">=3.0.3"
django-environ = ">=0.4.5"
django-graphql-jwt = ">=0.3.2"
django-versatileimagefield = ">=2.0.0"
whitenoise = ">=5.2.0"
celery = ">=5.1.2"
django-celery-results = ">=2.2.0"
django-celery-beat = ">=2.2.1"
requests = ">=2.25.1"
django-mptt = ">=0.11.0"
django-model-utils = ">=4.1.1"
django-prices = ">=2.1.1"
django-phonenumber-field = ">=4.0.0"
django-js-asset = ">=1.2.2"
django-statici18n = ">=2.2.0"
django-webpack-loader = ">=0.7.0"
django-money = ">=1.1.0"
django-templated-email = ">=3.0.0"
django-prices-openexchangerates = ">=1.0.0"
django-extensions = ">=3.1.3"
Babel = ">=2.9.1"
graphene-django-extras = ">=0.5.1"
django-allauth = ">=0.44.0"
drf-yasg = ">=1.20.0"
django-debug-toolbar = ">=3.2.2"
pytest-django = ">=4.4.0"
pytest-cov = ">=2.10.1"
factory-boy = ">=3.2.0"
bandit = ">=1.7.0"
safety = ">=1.10.3"
Pillow = ">=8.4.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2"